+++ spec/dummy/client/entries/app-kitten.js
@@ -64,6 +64,7 @@ import { TabBar } from 'kitten/components/navigation/tab-bar'
 // Notifications
 import { Alert } from 'kitten/components/notifications/alert'
 import { Badge } from 'kitten/components/notifications/badge'
+import { IconBadge } from 'kitten/components/notifications/icon-badge'
 
 // Popovers
 import { Popover } from 'kitten/components/popovers/popover'
@@ -154,6 +155,7 @@ import { KarlMakerWhoAmIStep } from 'app/maker/who-am-i/components/container'
 import { KarlAlert,
          KarlAlertWithButton,
          KarlAlertWithMarkdown } from 'kitten/karl/notifications/alert'
+import { KarlIconBadgeWithSvg } from 'kitten/karl/notifications/icon-badge'
 import { KarlLegoGrid } from 'kitten/karl/grid/lego-grid'
 
 ReactOnRails.register({
@@ -222,6 +224,7 @@ ReactOnRails.register({
   // Notifications
   Alert,
   Badge,
+  IconBadge,
 
   // Popovers
   CallToActionPopover,
@@ -266,6 +269,7 @@ ReactOnRails.register({
   KarlButtonAndText,
   KarlLineWithTooltip,
   KarlLineWithBigHeart,
+  KarlIconBadgeWithSvg,
   KarlImageExternalRichLink,
   KarlSvgExternalRichLink,
   KarlSimpleDropdown,