+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `IconBadge` component.
 - Feature: Add new `Badge` React component.
 - Feature: Add `LegoGrid` component.
 - Feature: Add `flexible` option to `k-grid-colSize` mixin.